Rename Dalli::Server into Dalli::Protocol::Binary
