+++ test/test_binary_protocol.rb
@@ -2,24 +2,24 @@
 
 require_relative "helper"
 
-describe Dalli::Server do
+describe Dalli::Protocol::Binary do
   describe "hostname parsing" do
     it "handles unix socket with no weight" do
-      s = Dalli::Server.new("/var/run/memcached/sock")
+      s = Dalli::Protocol::Binary.new("/var/run/memcached/sock")
       assert_equal "/var/run/memcached/sock", s.hostname
       assert_equal 1, s.weight
       assert_equal :unix, s.socket_type
     end
 
     it "handles unix socket with a weight" do
-      s = Dalli::Server.new("/var/run/memcached/sock:2")
+      s = Dalli::Protocol::Binary.new("/var/run/memcached/sock:2")
       assert_equal "/var/run/memcached/sock", s.hostname
       assert_equal 2, s.weight
       assert_equal :unix, s.socket_type
     end
 
     it "handles no port or weight" do
-      s = Dalli::Server.new("localhost")
+      s = Dalli::Protocol::Binary.new("localhost")
       assert_equal "localhost", s.hostname
       assert_equal 11211, s.port
       assert_equal 1, s.weight
@@ -27,7 +27,7 @@ describe Dalli::Server do
     end
 
     it "handles a port, but no weight" do
-      s = Dalli::Server.new("localhost:11212")
+      s = Dalli::Protocol::Binary.new("localhost:11212")
       assert_equal "localhost", s.hostname
       assert_equal 11212, s.port
       assert_equal 1, s.weight
@@ -35,7 +35,7 @@ describe Dalli::Server do
     end
 
     it "handles a port and a weight" do
-      s = Dalli::Server.new("localhost:11212:2")
+      s = Dalli::Protocol::Binary.new("localhost:11212:2")
       assert_equal "localhost", s.hostname
       assert_equal 11212, s.port
       assert_equal 2, s.weight
@@ -43,7 +43,7 @@ describe Dalli::Server do
     end
 
     it "handles ipv4 addresses" do
-      s = Dalli::Server.new("127.0.0.1")
+      s = Dalli::Protocol::Binary.new("127.0.0.1")
       assert_equal "127.0.0.1", s.hostname
       assert_equal 11211, s.port
       assert_equal 1, s.weight
@@ -51,7 +51,7 @@ describe Dalli::Server do
     end
 
     it "handles ipv6 addresses" do
-      s = Dalli::Server.new("[::1]")
+      s = Dalli::Protocol::Binary.new("[::1]")
       assert_equal "::1", s.hostname
       assert_equal 11211, s.port
       assert_equal 1, s.weight
@@ -59,7 +59,7 @@ describe Dalli::Server do
     end
 
     it "handles ipv6 addresses with port" do
-      s = Dalli::Server.new("[::1]:11212")
+      s = Dalli::Protocol::Binary.new("[::1]:11212")
       assert_equal "::1", s.hostname
       assert_equal 11212, s.port
       assert_equal 1, s.weight
@@ -67,7 +67,7 @@ describe Dalli::Server do
     end
 
     it "handles ipv6 addresses with port and weight" do
-      s = Dalli::Server.new("[::1]:11212:2")
+      s = Dalli::Protocol::Binary.new("[::1]:11212:2")
       assert_equal "::1", s.hostname
       assert_equal 11212, s.port
       assert_equal 2, s.weight
@@ -75,7 +75,7 @@ describe Dalli::Server do
     end
 
     it "handles a FQDN" do
-      s = Dalli::Server.new("my.fqdn.com")
+      s = Dalli::Protocol::Binary.new("my.fqdn.com")
       assert_equal "my.fqdn.com", s.hostname
       assert_equal 11211, s.port
       assert_equal 1, s.weight
@@ -83,7 +83,7 @@ describe Dalli::Server do
     end
 
     it "handles a FQDN with port and weight" do
-      s = Dalli::Server.new("my.fqdn.com:11212:2")
+      s = Dalli::Protocol::Binary.new("my.fqdn.com:11212:2")
       assert_equal "my.fqdn.com", s.hostname
       assert_equal 11212, s.port
       assert_equal 2, s.weight
@@ -91,26 +91,26 @@ describe Dalli::Server do
     end
 
     it "throws an exception if the hostname cannot be parsed" do
-      expect(lambda { Dalli::Server.new("[]") }).must_raise Dalli::DalliError
-      expect(lambda { Dalli::Server.new("my.fqdn.com:") }).must_raise Dalli::DalliError
-      expect(lambda { Dalli::Server.new("my.fqdn.com:11212,:2") }).must_raise Dalli::DalliError
-      expect(lambda { Dalli::Server.new("my.fqdn.com:11212:abc") }).must_raise Dalli::DalliError
+      expect(lambda { Dalli::Protocol::Binary.new("[]") }).must_raise Dalli::DalliError
+      expect(lambda { Dalli::Protocol::Binary.new("my.fqdn.com:") }).must_raise Dalli::DalliError
+      expect(lambda { Dalli::Protocol::Binary.new("my.fqdn.com:11212,:2") }).must_raise Dalli::DalliError
+      expect(lambda { Dalli::Protocol::Binary.new("my.fqdn.com:11212:abc") }).must_raise Dalli::DalliError
     end
   end
 
   describe "ttl translation" do
     it "does not translate ttls under 30 days" do
-      s = Dalli::Server.new("localhost")
+      s = Dalli::Protocol::Binary.new("localhost")
       assert_equal s.send(:sanitize_ttl, 30 * 24 * 60 * 60), 30 * 24 * 60 * 60
     end
 
     it "translates ttls over 30 days into timestamps" do
-      s = Dalli::Server.new("localhost")
+      s = Dalli::Protocol::Binary.new("localhost")
       assert_equal s.send(:sanitize_ttl, 30 * 24 * 60 * 60 + 1), Time.now.to_i + 30 * 24 * 60 * 60 + 1
     end
 
     it "does not translate ttls which are already timestamps" do
-      s = Dalli::Server.new("localhost")
+      s = Dalli::Protocol::Binary.new("localhost")
       timestamp_ttl = Time.now.to_i + 60
       assert_equal s.send(:sanitize_ttl, timestamp_ttl), timestamp_ttl
     end
@@ -118,7 +118,7 @@ describe Dalli::Server do
 
   describe "guard_max_value" do
     it "throws when size is over max" do
-      s = Dalli::Server.new("127.0.0.1")
+      s = Dalli::Protocol::Binary.new("127.0.0.1")
       value = OpenStruct.new(bytesize: 1_048_577)
 
       assert_raises Dalli::ValueOverMaxSize do
@@ -128,7 +128,7 @@ describe Dalli::Server do
   end
 
   describe "deserialize" do
-    subject { Dalli::Server.new("127.0.0.1") }
+    subject { Dalli::Protocol::Binary.new("127.0.0.1") }
 
     it "uses Marshal as default serializer" do
       assert_equal subject.serializer, Marshal
@@ -136,19 +136,19 @@ describe Dalli::Server do
 
     it "deserializes serialized value" do
       value = "some_value"
-      deserialized = subject.send(:deserialize, Marshal.dump(value), Dalli::Server::FLAG_SERIALIZED)
+      deserialized = subject.send(:deserialize, Marshal.dump(value), Dalli::Protocol::Binary::FLAG_SERIALIZED)
       assert_equal value, deserialized
     end
 
     it "raises UnmarshalError for broken data" do
       assert_raises Dalli::UnmarshalError do
-        subject.send(:deserialize, :not_marshaled_value, Dalli::Server::FLAG_SERIALIZED)
+        subject.send(:deserialize, :not_marshaled_value, Dalli::Protocol::Binary::FLAG_SERIALIZED)
       end
     end
 
     describe "custom serializer" do
       let(:serializer) { Minitest::Mock.new }
-      subject { Dalli::Server.new("127.0.0.1", serializer: serializer) }
+      subject { Dalli::Protocol::Binary.new("127.0.0.1", serializer: serializer) }
 
       it "uses custom serializer" do
         assert subject.serializer === serializer
@@ -159,7 +159,7 @@ describe Dalli::Server do
           raise NameError, "ddd"
         end
         assert_raises NameError do
-          subject.send(:deserialize, :some_value, Dalli::Server::FLAG_SERIALIZED)
+          subject.send(:deserialize, :some_value, Dalli::Protocol::Binary::FLAG_SERIALIZED)
         end
       end
 
@@ -168,7 +168,7 @@ describe Dalli::Server do
           raise NameError, "uninitialized constant Ddd"
         end
         assert_raises Dalli::UnmarshalError do
-          subject.send(:deserialize, :some_value, Dalli::Server::FLAG_SERIALIZED)
+          subject.send(:deserialize, :some_value, Dalli::Protocol::Binary::FLAG_SERIALIZED)
         end
       end
 
@@ -177,7 +177,7 @@ describe Dalli::Server do
           raise ArgumentError, "ddd"
         end
         assert_raises ArgumentError do
-          subject.send(:deserialize, :some_value, Dalli::Server::FLAG_SERIALIZED)
+          subject.send(:deserialize, :some_value, Dalli::Protocol::Binary::FLAG_SERIALIZED)
         end
       end
 
@@ -186,14 +186,14 @@ describe Dalli::Server do
           raise ArgumentError, "undefined class Ddd"
         end
         assert_raises Dalli::UnmarshalError do
-          subject.send(:deserialize, :some_value, Dalli::Server::FLAG_SERIALIZED)
+          subject.send(:deserialize, :some_value, Dalli::Protocol::Binary::FLAG_SERIALIZED)
         end
       end
     end
   end
 
   describe "multi_response_nonblock" do
-    subject { Dalli::Server.new("127.0.0.1") }
+    subject { Dalli::Protocol::Binary.new("127.0.0.1") }
 
     it "raises NetworkError when called before multi_response_start" do
       assert_raises Dalli::NetworkError do
